Implementado método findByEmail no UsuarioService, lógica de ativação transferida para o AtivacaoService e implementado serviço de renovação de código de ativação

+++ src/main/java/software/gabriel/easyjobs/service/AtivacaoService.java
@@ -9,13 +9,18 @@ import java.time.LocalDateTime;
 import java.util.Random;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import software.gabriel.easyjobs.dto.AtivacaoDTO;
+import software.gabriel.easyjobs.dto.UsuarioDTO;
 import software.gabriel.easyjobs.entity.Ativacao;
 import software.gabriel.easyjobs.entity.Usuario;
+import software.gabriel.easyjobs.exception.ativacao.AtivacaoInexistenteException;
 import software.gabriel.easyjobs.exception.ativacao.CodigoAtivacaoExpiradoException;
 import software.gabriel.easyjobs.exception.ativacao.CodigoAtivacaoIncorretoException;
+import software.gabriel.easyjobs.exception.ativacao.RenovacaoAntecipadaCodigoAtivacaoException;
 import software.gabriel.easyjobs.repository.AtivacaoRepository;
 import software.gabriel.easyjobs.service.mail.AtivacaoMail;
 import software.gabriel.easyjobs.service.mail.MailService;
+import software.gabriel.easyjobs.utils.AtivacaoUtils;
 
 /**
  *
@@ -28,6 +33,9 @@ public class AtivacaoService {
     AtivacaoRepository ativacaoRepository;
 
     @Autowired
+    UsuarioService usuarioService;
+
+    @Autowired
     MailService mailService;
 
     public void cadastrar(Usuario usuario) {
@@ -36,34 +44,54 @@ public class AtivacaoService {
         enviarEmail(ativacao);
     }
 
-    public void ativar(Usuario usuario, String codigo) {
-        Ativacao ativacao = ativacaoRepository.findByUsuario(usuario);
+    public void ativar(AtivacaoDTO ativacaoDTO) {
+        Usuario usuario = usuarioService.findByEmail(ativacaoDTO.getEmail());
+        Ativacao ativacao = validarExistenciaAtivacao(usuario);
         if (LocalDateTime.now().isAfter(ativacao.getExpiracao())) {
             renovarCodigo(ativacao, usuario);
             throw new CodigoAtivacaoExpiradoException();
         }
-        if (codigo.equals(ativacao.getCodigo())) {
+        if (ativacaoDTO.getCodigo().equals(ativacao.getCodigo())) {
             ativacaoRepository.delete(ativacao);
+            usuarioService.ativar(usuario);
         } else {
             throw new CodigoAtivacaoIncorretoException();
         }
+    }
+
+    public void renovar(UsuarioDTO usuarioDTO) {
+        Usuario usuario = usuarioService.findByEmail(usuarioDTO.getEmail());
+        Ativacao ativacao = validarExistenciaAtivacao(usuario);
+        renovarCodigo(ativacao, usuario);
     }
 
     private Ativacao gerarCodigo(Ativacao ativacao) {
         Random secureRandom = new SecureRandom();
         StringBuilder sb = new StringBuilder();
 
-        for (int i = 0; i < 6; i++) {
+        for (int i = 0; i < AtivacaoUtils.TAMANHO_CODIGO_ATIVACAO; i++) {
             sb.append(secureRandom.nextInt(9));
         }
 
         ativacao.setCodigo(sb.toString());
-        ativacao.setExpiracao(LocalDateTime.now().plusDays(1));
+        ativacao.setGeracao(LocalDateTime.now());
+        ativacao.setExpiracao(LocalDateTime.now().plusDays(AtivacaoUtils.DIAS_EXPIRACAO_CODIGO_ATIVACAO));
 
         return ativacao;
     }
-    
+
+    private Ativacao validarExistenciaAtivacao(Usuario usuario) {
+        Ativacao ativacao = ativacaoRepository.findByUsuario(usuario);
+        if (ativacao == null) {
+            throw new AtivacaoInexistenteException();
+        }
+        return ativacao;
+    }
+
     private void renovarCodigo(Ativacao ativacao, Usuario usuario) {
+        if (LocalDateTime.now().isBefore(ativacao.getGeracao().plusMinutes(AtivacaoUtils.MINUTOS_RENOVACAO_CODIGO_ATIVACAO))) {
+            throw new RenovacaoAntecipadaCodigoAtivacaoException();
+        }
         ativacaoRepository.delete(ativacao);
         cadastrar(usuario);
     }
